refactor!: retire abicheck.cli's lazy compatibility-alias surface
`abicheck/cli.py` carried a name-to-owner table of 82 private helpers that
had physically moved to sibling modules (`frontends/cli/moved.py`), resolved
at attribute access by a module-level `__getattr__`, plus a custom module
class that raised on assignment to any of those names.

That guard was not decoration. A `monkeypatch.setattr` against a lazily
resolved alias records the resolved original and re-assigns it on undo,
permanently shadowing the lookup -- so every later caller reads a stale
function and every later test that patches the true owner is silently
ignored. It had already landed as an order-dependent CI failure two test
files away from the one that caused it. The shape was therefore: move the
implementation, keep the old import path, add a lazy resolver to serve it,
then add an assignment guard to stop the resolver being poisoned.

All 109 call sites now import from the owning module -- 9 production modules
and 25 test files -- so the table, the resolver and the guard are deleted
together rather than one being left to justify the next.

Two of those call sites are worth naming. `cli_buildsource.py` reached
`_missing_requested_evidence_layers`/`_classify_missing_layers` -- both
defined in that same file, forty lines above -- through `abicheck.cli`,
purely so a monkeypatch on the facade would land; it now calls them
directly. `cli_compare_helpers.py` carried a comment claiming its
`_normalize_binary_input` call went through `cli` for the same reason, which
it did not: the comment outlived the indirection it described.

Tests: `tests/test_cli_moved_surface.py` existed to prove the aliases still
resolved, and is retired with them. Its one behavioral claim -- ADR-061
Phase 4's acceptance criterion, `__all__ == ["main"]` and under 150 lines,
deliberately pinned in a test rather than `modules.yaml`'s `facades` rule
(which no module defining a Click group can satisfy) -- moves to
`tests/test_cli_root_surface.py`, which already owns the root module's
executable contract.

Architecture baselines: two files grew by exactly one line, because they
imported names from `abicheck.cli` that are owned by *different* modules, so
one facade import becomes two. That is the mechanical cost of removing an
indirection layer, not new responsibility -- and the import count is now the
file's real dependency count rather than one the table was hiding.
`cli_project.py` would have crossed the 800-line production cap the same
way; rather than open a debt entry for an import split, an over-split import
in the same block (held multi-line only by a magic trailing comma, and 87
characters when joined) was collapsed, leaving the file at 798.

Verified: `mypy abicheck/` clean over 794 files; `ruff` clean; architecture
gate 0 errors; ai-readiness 0 errors; 1496 tests pass across every migrated
test file plus `test_cli_contract.py`; `getattr(cli, "_safe_write_output")`
now raises the ordinary AttributeError.

BREAKING CHANGE: `abicheck.cli` no longer re-exports the ~80 private helpers
it once defined. They were private (underscore-prefixed) and are unchanged
at their owning modules; import from the owner, and patch the owner.
