+++ tests/test_cli_unit.py
@@ -687,7 +687,7 @@ class TestNoFailOnAdditionsFlag:
         # Unrecognised option → Click usage error, remapped to the dedicated
         # usage-error code (outside the compare result space {0,1,2,4}) so it is
         # not mistaken for a "2 = source break" verdict.
-        from abicheck.cli import _EXIT_USAGE_ERROR
+        from abicheck.frontends.cli.runtime import _EXIT_USAGE_ERROR
         assert result.exit_code == _EXIT_USAGE_ERROR
